requirements.txt
spacy>=3.8.0
PyMuPDF>=1.21.0
python-docx>=0.8.11
streamlit>=1.20.0
pandas>=1.5.0
reportlab>=3.6.12
pytest>=7.0.0

# Optional / dev
# You can add packages here for linting or CI caching if needed (e.g. ruff, black)